Start working on Navigation Routes now. Implement the efficient way of navigating instead of the traditional.
