+++ React-06/comps/src/App.js
@@ -1,21 +1,8 @@
-import { useState } from "react";
-import Dropdown from "./components/Dropdown";
+
 
 function App(){
-
-    const [selection, setSelection] = useState(null);
-
-    const handleSelect = (option) =>{
-        setSelection(option);
-    };
-
-    const options = [
-        {label : 'Red', value: 'red'},
-        {label : 'Blue', value: 'blue'},
-        {label: 'Green', value: 'green'}
-    ]
-
-    return <Dropdown options = {options} value ={selection} onChange = {handleSelect}/>;
+    return <div>
+        </div>
 }
 
 export default App;
