+++ React-06/comps/src/components/Dropdown.js
@@ -1,76 +1,24 @@
-import { useState, useEffect, useRef } from "react";
+import { useState } from "react";
+import Dropdown from "../components/Dropdown";
 
-// import handleSelect from "../"
-import {GoChevronDown} from "react-icons/go";
+function DropdownPage(){
 
+    const [selection, setSelection] = useState(null);
 
-function Dropdown({options, value, onChange}){
+    const handleSelect = (option) =>{
+        setSelection(option);
+    };
 
-    const [isOpen, setIsOpen] = useState(false);
-
-
-    const [selected, setSelected] = useState('Select...');
-
-    const divEl = useRef();
-
-
-    useEffect(()=>{
-        const handler = (event) =>{
-            // console.log(divEl);
-            if(!divEl.current){
-                return;
-            }
-
-            if(divEl.current.contains(event.target)){
-                setIsOpen(false);
-            }
-        }
-
-        document.addEventListener('click', handler, true);
-
-        return ()=>{
-            document.removeEventListener('click', handler);
-        }
-    }, []);
-
-
-    const handleClick = () =>{
-        if(isOpen){
-            setIsOpen(false);          
-        }
-        else{
-            setIsOpen(true);
-        }
-    }
-
-    const handleOptionClick = (option) =>{
-        setIsOpen(false);
-
-        setSelected(option.label);
-
-        onChange(option);
-    }
-
-    // renderedOptions below is an array of 
-    // div elements consiting of all option.labels.
-    const renderedOptions = options.map((option) => {
-        return <div className = "hover:bg-sky-100 rounded cursor-pointer p-1" key={option.value} onClick={()=>handleOptionClick(option) }>
-            {option.label}
-        </div>
-    });
-
+    const options = [
+        {label : 'Red', value: 'red'},
+        {label : 'Blue', value: 'blue'},
+        {label: 'Green', value: 'green'}
+    ]
 
     return (
-    <div ref={divEl} className="w-48 relative ">
-        <div 
-            className="flex justify-between items-center cursor-pointer rounded p-3 shadow bg-white w-full" 
-            onClick={handleClick}>
-                {value}
-            <GoChevronDown className="text-lg"/>
-        </div>
-        {isOpen && <div className="absolute top-full border rounded p-3 shadow bg-white w-full">{renderedOptions}</div>}
+    <div className="flex">
+        <Dropdown options = {options} value ={selection} onChange = {handleSelect}/>
     </div>);
 }
 
-
-export default Dropdown;
+export default DropdownPage;
